+++ secretspec-ex/config/config.exs
@@ -1,5 +1,1 @@
 import Config
-
-config :secretspec, SecretSpec.Native,
-  crate: "secretspec_native",
-  mode: :release